Change return type of #load to void, simplify
ConfigurationSource#load now returns void instead of a boolean
indicating success. If load fails a ConfigurationException is now
thrown. This makes figuring out why the configuration won't load a lot
easier, as the whole stack trace is present.

In the process of refactoring, usage of Apache Commons Configuration's
ConfigurationException was dropped and replaced with this project's own
ConfigurationException.