Store modification time in config
Modifications on ConcurrentConfiguration instances cause a parameter to
be written there containing a modification timestamp. A new helper class
ConfigurationMonitor can be used to query this parameter to determine if
the configuration was updated since the last time the monitor was
queried.

Useful if computationally intensive tasks are performed on configuration
data and subsequently cached.